added libs directory for angular libraries - created a directory to hold all libraries for the angular app - moved configuration and shared to new libs directory - updated all paths
