+++ src/AngularClient/ClientApp/src/app/app-routing.module.ts
@@ -1,12 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AuthGuardService } from './shared/services/auth-guard.service';
+import { AuthGuardService } from 'libs/shared';
 
 const routes: Routes = [
   {
         path: '',
         canActivate: [AuthGuardService],
-        canLoad: [AuthGuardService],
         children: [
             { path: '', loadChildren: () => import('./home/home.module').then(m => m.HomeModule) }
         ]
