+++ src/AngularClient/ClientApp/src/app/app.module.ts
@@ -1,10 +1,10 @@
-﻿import { BrowserModule } from '@angular/platform-browser';
+import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, APP_INITIALIZER } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 
-import { SharedModule } from './shared';
-import { ConfigurationModule } from './configuration';
+import { SharedModule } from 'libs/shared';
+import { ConfigurationModule } from 'libs/configuration';
 import { AuthenticationModule } from './authentication';
 
 import { AppRoutingModule } from './app-routing.module';
