+++ src/AngularClient/ClientApp/src/app/home/home.module.ts
@@ -1,7 +1,7 @@
-﻿import { NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { AuthenticationModule } from 'authentication';
+import { AuthenticationModule } from '../authentication';
 
 import { HomeRoutingModule } from './home-routing.module';
 import { HomeComponent } from './home.component';
