+++ src/AngularClient/ClientApp/src/libs/configuration/app.config.ts
