+++ src/AngularClient/ClientApp/src/libs/configuration/configuration.module.ts
