+++ src/AngularClient/ClientApp/src/libs/configuration/configuration.service.ts
