+++ src/AngularClient/ClientApp/src/libs/configuration/exports.ts
