+++ src/AngularClient/ClientApp/src/libs/configuration/index.ts
