+++ src/AngularClient/ClientApp/src/libs/shared/index.ts
