+++ src/AngularClient/ClientApp/src/libs/shared/services/auth-guard.service.ts
