+++ src/AngularClient/ClientApp/src/libs/shared/shared.module.ts
